+++ NEWS.md
@@ -1,3 +1,5 @@
+# pipapi (development version)
+
 # pipapi 1.2.0
 
 ## New features