+++ NEWS.md
@@ -1,5 +1,9 @@
 # pipapi (development version)
 
+## Bug fixes
+- Hot fix to handle failing of `create_vector_countries()` when `country="ALL"`
+- Fix cp-download
+
 # pipapi 1.2.0
 
 ## New features